requirement.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
idna==2.10
pycparser==2.20
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
tabulate==0.8.7
tornado==6.0.4
urllib3==1.25.10
